lab2/producer/pyproject.toml
[tool.poetry]
name = "bigdata-labs-producer"
version = "0.1.0"
description = ""
authors = ["Blackgolyb <andrejomelnickij@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
websockets = "^11.0.3"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
